feat: add module func refactor
